[WIP]Updated .gitignore file.
