+++ bmi-calculator-flutter/lib/constants.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+const kBottomContainerHeight = 70.0;
+const bottomContainerColor = Color(0xFFEB1555);
+const inactiveColor = Color(0xFF111328);
+const activeColor = Color(0xFF1D1E33);
+const anotherColor = Color(0xFF8D8E98);
+const defaultColor = Color(0xFF0E0C20);
+const spaceBetween = 10.0;
+const iconSize = 70.0;
+const labelSize = 20.0;
+
+const labelTextStyle = TextStyle(
+  fontSize:  labelSize,
+  color: anotherColor
+);
